pyproject.toml
[tool.poetry]
name = "python-sucks"
version = "0.1.0"
description = ""
authors = ["storvus <anton.scherbatov@invitae.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
django-ckeditor = "^6.7.0"
django-taggit = "^5.0.1"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
